add listeners in declorate style
