--- src/blocks/popup/_is-opened/popup_is-opened.css
@@ -1,3 +1,0 @@
-.popup.popup_is-opened {
-  display: flex;
-}
