+++ src/blocks/popup/popup.css
@@ -4,7 +4,6 @@
 @import url(__form/popup__form.css);
 @import url(__input/popup__input.css);
 @import url(__button/popup__button.css);
-@import url(_is-opened/popup_is-opened.css);
 @import url(__image/popup__image.css);
 @import url(__container/popup__container.css);
 
@@ -18,5 +17,4 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  display: none;
 }
